split off fast path and settle call out of run loop
